feat(core): passing api to manager
